# Conflicts:
#	bsp/boards/radio.h
#	openstack/02a-MAClow/IEEE802154E.c

+++ bsp/boards/radio.h
@@ -67,7 +67,7 @@ void                radio_rfOn(void);
 void                radio_rfOff(void);
 // TX
 void                radio_loadPacket_prepare(uint8_t* packet, uint8_t len);
-void                radio_loadPacket(uint8_t* packet, uint8_t len);
+void                radio_loadPacket(uint8_t* packet, uint16_t len);
 void                radio_txEnable(void);
 void                radio_txNow(void);
 // RX
